Add Copy image to repo
Not yet using this for anything, but I might try to add it on to the button.

I was also originally looking at rearranging some of the UI in the BulkPropertyItem, but I ended up not liking how it turned out. Soooo we're going to stick with what we have.